diagnostics: Fix incorrect integer types in memory routine v2
base::File use int64_t as the return type of File::GetLength. Use the
same type accordingly.

Change the local variable that stores the return value of File::Read
from size_t to int since the return value could be negative when the
read fails.

BUG=b:281940761
TEST=USE="-cros-debug" FEATURES=test emerge-volteer diagnostics

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4531666
Tested-by: Wei-Luan Wang <weiluanwang@google.com>
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>

+++ diagnostics/cros_healthd/routines/memory_and_cpu/memory_v2.h
@@ -83,7 +83,7 @@ class MemoryRoutineV2 final : public BaseRoutineControl {
   // output capturing.
   base::File stdout_file_;
   // Stores the number of bytes the stdout file has been read so far.
-  size_t read_stdout_size_;
+  int64_t read_stdout_size_;
   // Stores the parsed stdout result.
   std::vector<std::vector<std::string>> parsed_memtester_result_;
   // Stores the number of kib the memtester should test for as requested by the
